+++ src/wetterdienst/provider/rmi/observation/__init__.py
@@ -1,0 +1,9 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from wetterdienst.provider.rmi.observation.api import RmiObservationRequest
+from wetterdienst.provider.rmi.observation.metadata import RmiObservationMetadata
+
+__all__ = [
+    "RmiObservationMetadata",
+    "RmiObservationRequest",
+]